refactor(cdk/schematics): use schematics addDependency helper in cdk ng-add
The `addDependency` helper from `@schematics/angular/utility` reduces the
amount of custom code to conditionally add the `@angular/cdk` dependency
within the package's `ng add` schematic. This also allows remove of some
of the additional helper code within the schematic. Unfortunately, not all
can be removed as some is used within the unit tests for the schematic.
